material composer file reorganization
