+++ frontend/dialob-composer-material/src/components/translations/LanguageEditor.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Box } from '@mui/material';
+
+const LanguageEditor: React.FC = () => {
+  return (
+    <Box>
+      LanguageEditor
+    </Box>
+  );
+};
+
+export default LanguageEditor;
