+++ frontend/dialob-composer-material/src/components/translations/MissingTranslations.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Box } from '@mui/material';
+
+const MissingTranslations: React.FC = () => {
+  return (
+    <Box>
+      MissingTranslations
+    </Box>
+  );
+};
+
+export default MissingTranslations;
