+++ frontend/dialob-composer-material/src/components/translations/TranslationFileEditor.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Box } from '@mui/material';
+
+const TranslationFileEditor: React.FC = () => {
+  return (
+    <Box>
+      TranslationFileEditor
+    </Box>
+  );
+};
+
+export default TranslationFileEditor;
