+++ frontend/dialob-composer-material/src/components/tree/NavigationTreeItem.tsx
